bt-wlan-coex: Proper termination of BTC module without relying on SIGTERM
If a service is stopped via SIGKILL, this abruptly stops the service thereby
not giving chance for the service to complete its required cleanup. BTC module
now handles this graceful cleanup without relying on SIGTERM. This also
ensures that the missing BT off required by WLAN is posted by BTC when
Bluetooth is turned off.

CRs-fixed: 326323
Change-Id: I117a3c59ce14b3fcf4d3c9ec3a6ade5341b5013a